Update cqlsh files to Cassandra 5.0.0
